[CI] Using sphinx-build as an alternative to make
Gitlab seems to be having issues with make, which does not recognise the
sphinx-build command from the PATH. This might be a temporary problem as the
same script was working fine a month ago. Using sphinx-build directly might
solve the issue for now and could remain a permanent fix if the CI succeeds.